2015-08-21 [] === (>|) The Wizward works now. cant find any more errors (>>) changed the directory path readings to the install4j runtime api

+++ src/entity/info/MREPrevalenceSheets.java
@@ -148,7 +148,7 @@ public class MREPrevalenceSheets {
         mapBedsInUse = new HashMap<>();
 
         for (Homes home : HomesTools.getAll()) {
-            short maxLevel = RoomsTools.getMaxLevel(home);
+            int maxLevel = RoomsTools.getMaxLevel(home);
             mapBedsTotal.put(home, new int[maxLevel + 1]);
             mapBedsInUse.put(home, new int[maxLevel + 1]);
